Move authentication functions to authRouter
