+++ src/controllers/authController.js
@@ -14,6 +14,7 @@ import { generateToken } from '../utils/generateJwtToken.js';
 // Main Controller functions
 // write description for each function
 
+
 export const registerUser = async (req, res) => {
     const { email, whatsapp, otpMethod = 'whatsapp' } = req.body;
 
@@ -235,27 +236,10 @@ export const forgotPassword = async (req, res) => {
     }
 };
 
-export const resetPassword = async (req, res) => {
-    const { token, newPassword } = req.body;
-
-    try {
-        const user = await CompleteUser.findOne({
-            resetPasswordToken: token,
-            resetPasswordExpires: { $gt: Date.now() }
-        });
-
-        if (!user) {
-            return res.status(400).json({ error: 'Invalid or expired reset token' });
-        }
-
-        const hashedPassword = await bcrypt.hash(newPassword, 10);
-        user.password = hashedPassword;
-        user.resetPasswordToken = undefined;
-        user.resetPasswordExpires = undefined;
-        await user.save();
-
-        res.status(200).json({ message: 'Password reset successful' });
-    } catch (error) {
-        res.status(400).json({ error: 'Password reset failed' });
+// refresh access token (still needs to be completed)
+export const refreshAccessToken = async (req, res) => {
+    const { refreshToken } = req.body;
+    if (!refreshToken) {
+        return res.status(400).json({ error: 'Refresh token required' });
     }
-};
+}
